Module Alcotest.Core

exception Check_error of string

module IntSet = Core.IntSet

module type S = sig ... end

module Make = Core.Make

Functor for building a tester that sequences tests of type ('a -> unit M.t) within a given concurrency monad M.t. The run and run_with_args functions must be scheduled in a global event loop. Intended for use by the Alcotest_lwt and Alcotest_async backends.