Module Alcotest_engine

Alcotest_engine provides a platform-independent test framework.

The main building blocks and combinators are defined here. These can be used to defined tests. The platform-specific runners for these tests are in alcotest, alcotest-lwt, alcotest-async and alcotest-mirage.

module V1 : sig ... end

Version 1 of the user-facing Alcotest API.

module Monad : sig ... end

Monad signatures for use with V1.Core and V1.Cli.

module Platform : sig ... end

Defines platform-dependent functions.

module Private : sig ... end

These modules are exposed for use internally by other Alcotest packages. They do not provide a stable interface.

module Formatters : sig ... end

Alcotest uses Format.std_formatter and Format.err_formatter formatters. However, in a parallel context (OCaml 5 and domains), using these values in parallel can lead to data-races, since these values are not domains-safe. As such, Alcotest offers a way to create your own formatter equivalent in behavior to Format.std_formatter and Format.err_formatter (i.e. they write well on 1 and 2) but they can be used without risk even if another library (such as Logs) uses Format.std_formatter and/or Format.err_formatter and is used in parallel.