Module Alcotest_engine.V1

Version 1 of the user-facing Alcotest API.

Assert functions

module Test : sig ... end

Monadic test runners

These modules provide the ability to run tests inside a concurrency monad: that is, to sequence test cases of type 'a -> unit m into a computation of type 'a -> unit m (for some concurrency monad m) with can then be scheduled in a main event loop. For tests using Lwt.t or Async_kernel.Deferred.t, use the Alcotest_lwt and Alcotest_async packages directly.

module Core : sig ... end

Defines monadic test runners without command-line interfaces.

module Cli : sig ... end

Wraps Core to provide a command-line interface.