Module Model.Test_name

type t

val v : name:string -> index:int -> t

val name : t -> string

val index : t -> int

val pp : t Fmt.t

Pretty-print the unescaped test-case name

val file : t -> string

An escaped form of the test name with .output suffix.

val length : t -> int

The approximate number of terminal columns consumed by pp_name.

val compare : t -> t -> int

Order lexicographically by name, then by index.