Module type Alcotest_engine__Cli.S

type return

The return type of each test case run by Alcotest. For the standard Alcotest module, return = unit. The concurrent backends Alcotest_lwt and Alcotest_async set return = unit Lwt.t and return = Async_kernel.Deferred.t respectively.

type speed_level = [

| `Quick

| `Slow

]

Speed level of a test. Tests marked as `Quick are always run. Tests marked as `Slow are skipped when the `-q` flag is passed.

type 'a test_case = string * speed_level * ('a -> return)

A test case is a UTF-8 encoded documentation string, a speed level and a function to execute. Typically, the testing function calls the helper functions provided below (such as check and fail).

exception Test_error

The exception return by run in case of errors.

val test_case : string -> speed_level -> ('a -> return) -> 'a test_case

test_case n s f is the test case n running at speed s using the function f.

type 'a test = string * 'a test_case list

A test is a UTF-8 encoded name and a list of test cases. The name can be used for filtering which tests to run on the CLI.

val list_tests : 'a test list -> return

Print all of the test cases in a human-readable form

type 'a with_options = ?⁠and_exit:bool -> ?⁠verbose:bool -> ?⁠compact:bool -> ?⁠tail_errors:[ `Unlimited | `Limit of int ] -> ?⁠quick_only:bool -> ?⁠show_errors:bool -> ?⁠json:bool -> ?⁠filter:(Re.re option * int list option) -> ?⁠log_dir:string -> ?⁠bail:bool -> 'a

The various options taken by the tests runners run and run_with_args:

and_exit (default true). Once the tests have completed, exit with return code 0 if all tests passed, otherwise 1.

verbose (default false). Display the test std.out and std.err (rather than redirecting to a log file).

compact (default false). Compact the output of the tests.

tail_errors (default unlimited). Show only the last N lines of output of failed tests.

quick_only (default false). Don't run tests with the `Slow speed level.

show_errors (default false). Display the test errors.

json (default false). Print test results in a JSON-compatible format.

log_dir (default "$PWD/_build/_tests/"). The directory in which to log the output of the tests (if verbose is not set).

bail (default false). If true, stop running the tests after the first failure.

val run : (string -> unit test list -> return) with_options

val run_with_args : (string -> 'a -> 'a test list -> return) with_options

val run : (?⁠argv:string array -> string -> unit test list -> return) with_options

run n t runs the test suite t. n is the name of the tested library.

The optional argument and_exit controls what happens when the function ends. By default, and_exit is set, which makes the function exit with 0 if everything is fine or 1 if there is an issue. If and_exit is false, then the function raises Test_error on error.

The optional argument argv specifies command line arguments sent to alcotest like "--json", "--verbose", etc. Note that this array will be treated like a regular Sys.argv, so the array must have at least one element, and the first element will be treated as if it was the command name and thus ignored for the purposes of option processing. So ~argv:[||] is an error, ~argv:[| "--verbose" |] will have no effect, and ~argv:[| "ignored"; "--verbose" |] will successfully pass the verbose option.

val run_with_args : (?⁠argv:string array -> string -> 'a Cmdliner.Term.t -> 'a test list -> return) with_options

run_with_args n a t Similar to run a t but take an extra argument a. Every test function will receive as argument the evaluation of the Cmdliner term a: this is useful to configure the test behaviors using the CLI.