Module Alcotest_engine__Core

module type S = Alcotest_engine__.Core_intf.S

module type MAKER = Alcotest_engine__.Core_intf.MAKER

exception Check_error of unit Fmt.t

module Make : MAKER

Functor for building a tester that sequences tests of type ('a -> unit M.t) within a given concurrency monad M.t. The run and run_with_args functions must be scheduled in a global event loop. Intended for use by the Alcotest_lwt and Alcotest_async backends.