Module Alcotest_engine

Alcotest_engine provides a platform-independent test framework.

The main building blocks and combinators are defined here. These can be used to defined tests. The platform-specific runners for these tests are in alcotest, alcotest-lwt, alcotest-async and alcotest-mirage.

module V1 : sig ... end

Version 1 of the user-facing Alcotest API.

module Monad : sig ... end

Monad signatures for use with V1.Core and V1.Cli.

module Platform : sig ... end

Defines platform-dependent functions.

module Private : sig ... end

These modules are exposed for use internally by other Alcotest packages. They do not provide a stable interface.