Module Alcotest_engine

Alcotest_engine provides a platform-independent test framework.

The main building blocks and combinators are defined here. These can be used to defined tests. The platform-specific runners for these tests are in alcotest, alcotest-lwt, alcotest-async and alcotest-mirage.

Assert functions

module Test : sig ... end

Monadic test runners

module Core : sig ... end

Defines monadic test runners without command-line interfaces.

module Cli : sig ... end

Wraps Core to provide a command-line interface.

module Monad : sig ... end

Monad signatures for use with Core and Cli.

module Platform : sig ... end

Defines platform-dependent functions.

module Private : sig ... end

These modules are exposed for use internally by other Alcotest packages. They do not provide a stable interface.