Module Alcotest_engine.Core

Defines monadic test runners without command-line interfaces.

exception Check_error of unit Fmt.t

module IntSet : Stdlib.Set.S with type IntSet.elt = int

module type S = sig ... end

module type MAKER = functor (P : Platform.MAKER) -> functor (M : Monad.S) -> S with type return = unit M.t

module Make : MAKER

Functor for building a tester that sequences tests of type ('a -> unit M.t) within a given concurrency monad M.t. The run and run_with_args functions must be scheduled in a global event loop. Intended for use by the Alcotest_lwt and Alcotest_async backends.