Module V1.Core

Defines monadic test runners without command-line interfaces.

module type S = sig ... end

module type MAKER = sig ... end

module Make : MAKER

Functor for building a tester that sequences tests of type ('a -> unit M.t) within a given concurrency monad M.t. The run and run_with_args functions must be scheduled in a global event loop. Intended for use by the Alcotest_lwt and Alcotest_async backends.